Add RSA key management and face authentication services
- Implemented KeyIdGenerator for generating key IDs from public keys.
- Created PemKeyParse for parsing PEM encoded private and public keys.
- Developed RsaKeyGenerator for generating RSA key pairs.
- Introduced RsaKeyManagerService for managing RSA keys, including loading from files and generating new keys if not found.
- Added RsaKeyProvider interface for providing RSA keys.
- Implemented CookiesService for managing secure cookies for access and refresh tokens.
- Developed UserEnrollService for user enrollment and verification using face authentication.
- Created JwksService for providing JSON Web Key Set (JWKS) for public keys.
- Implemented OAuth2AuthenticationService for handling OAuth2 authentication flows.
- Added OpenIdConfigurationService for providing OpenID configuration.
- Created UserInforService for retrieving user information based on access tokens.
- Updated application properties for JWT and cookie configurations.
- Removed face-login.html template as part of the refactor.

+++ src/main/java/com/Authentication/AuthService/dto/RefreshTokenResponseDto.java
@@ -20,7 +20,7 @@ public class RefreshTokenResponseDto {
     private String tokenType = "Bearer";
 
     @JsonProperty("expires_in")
-    private Long expiresIn;
+    private int expiresIn;
 
     @JsonProperty("refresh_token")
     private String refreshToken;
